# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for Osau AMS.",
        "title": "Osau AMS API",
        "contact": {
            "name": "API Support",
            "email": "bakit.german.work@icloud.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/discipline_types": {
            "get": {
                "description": "Get a list of all discipline types",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DisciplineTypes"
                ],
                "summary": "Get all discipline types",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.DisciplineType"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing discipline type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DisciplineTypes"
                ],
                "summary": "Update a discipline type",
                "parameters": [
                    {
                        "description": "Discipline type info",
                        "name": "discipline_type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PutDisciplineTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new discipline type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DisciplineTypes"
                ],
                "summary": "Create a discipline type",
                "parameters": [
                    {
                        "description": "Discipline type info",
                        "name": "discipline_type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateDisciplineTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Partially update an existing discipline type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DisciplineTypes"
                ],
                "summary": "Partially update a discipline type",
                "parameters": [
                    {
                        "description": "Discipline type info",
                        "name": "discipline_type",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PatchDisciplineTypeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/discipline_types/{id}": {
            "get": {
                "description": "Get a discipline type by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DisciplineTypes"
                ],
                "summary": "Get a discipline type by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Discipline type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.DisciplineType"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an existing discipline type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DisciplineTypes"
                ],
                "summary": "Delete a discipline type",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Discipline type ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.DisciplineType": {
            "type": "object",
            "properties": {
                "discipline_type_id": {
                    "type": "integer"
                },
                "discipline_type_name": {
                    "type": "string"
                }
            }
        },
...
        "handler.CreateDisciplineTypeRequest": {
            "type": "object",
            "required": [
                "discipline_type_name"
            ],
            "properties": {
                "discipline_type_name": {
                    "type": "string"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handler.PatchDisciplineTypeRequest": {
            "type": "object",
            "required": [
                "discipline_type_id"
            ],
            "properties": {
                "discipline_type_id": {
                    "type": "integer"
                },
                "discipline_type_name": {
                    "type": "string"
                }
            }
        },
...
        "handler.PutDisciplineTypeRequest": {
            "type": "object",
            "required": [
                "discipline_type_id",
                "discipline_type_name"
            ],
            "properties": {
                "discipline_type_id": {
                    "type": "integer"
                },
                "discipline_type_name": {
                    "type": "string"
                }
            }
        },
...
        "handler.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}